requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
bcrypt==4.0.1
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.135.3
greenlet==3.3.2
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dotenv==1.2.2
python-multipart==0.0.22
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.42.0
